38G1FH7 split the scrubber's parts into the row, the track and the canvas
ScrubberParts was 1900 lines of three unrelated things: the control row above
the chart, the chart's own drawn pieces, and the canvas the punchcard is
painted to. Each is now its own file, importing only what it draws with.

ScatterDot goes with it. The canvas took over every point it used to render
and nothing has mounted it since.
