+++ source/gui/client/components/ScrubberControls.tsx
@@ -1,43 +1,24 @@
-// Every piece of the time scrubber's markup. Presentational only: props in,
-// JSX out, no state but the needle's own hover. TimeScrubber owns the data and
-// arranges these; the maths they draw against lives in lib/scrubber.
+// The row of controls above the chart: what the window covers, which series
+// are drawn, where the needle steps to next, and the button that plays the
+// whole thing back. Presentational — props in, JSX out; the state behind
+// them is TimeScrubber's.
 
-import {memo, useCallback, useEffect, useMemo, useRef, useState} from 'react';
+import {useEffect, useRef, useState} from 'react';
 import {GUI_THEME} from '../lib/gui-theme';
 import {
-	barGrowAnimation,
-	BAR_ENTRANCE_TOTAL_MS,
-	barWidthCss,
-	BUCKET_HIGHLIGHT_COLOR,
-	clamp,
-	DOT_EXIT_TOTAL_MS,
-	dotEntranceScale,
-	dotExitScale,
-	EVENTS_MODE_VERTICAL_PADDING,
-	EVENTS_SCATTER_HEIGHT,
-	FADE_IN_ANIMATION,
 	formatPeriodLabel,
-	HOVER_HINT_WIDTH,
 	isPeriodWindow,
 	LayoutMode,
-	NEEDLE_COLOR,
-	NEEDLE_GRIP_WIDTH,
 	PeriodRange,
-	RANGE_SELECTION_COLOR,
-	RANGE_SELECTION_EDGE,
 	Scope,
 	scopeButtonLabel,
 	SCOPES,
-	Segment,
 	BOARD_VIEWS,
 	boardViewColor,
 	soleVisibleIdentity,
 	identityAxisFor,
 	BoardView,
 	EventCategory,
-	SEGMENT_HIGHLIGHT_COLOR,
-	TRACK_HEIGHT,
-	VolumeBar,
 } from '../lib/scrubber';
 import {GuiEventIdentity} from '../lib/gui-state.model';
 import {Checkbox} from './Checkbox';
@@ -1141,781 +1122,3 @@ export const ScrubberPlayButton = ({
 // Deliberately one block spanning both charts and the gap: hovering a commit
 // lights up the same day in the issue track above, which is what makes the two
 // halves read as one time grid.
-export const SegmentHighlight = ({
-	segment,
-	fractionForTime,
-}: {
-	segment: Segment;
-	fractionForTime: (time: number) => number;
-}) => {
-	const left = fractionForTime(segment.start);
-
-	return (
-		<div
-			style={{
-				position: 'absolute',
-				left: `${left * 100}%`,
-				width: `${(fractionForTime(segment.end) - left) * 100}%`,
-				top: 0,
-				bottom: 0,
-				background: SEGMENT_HIGHLIGHT_COLOR,
-				pointerEvents: 'none',
-				display: 'flex',
-				justifyContent: 'center',
-				paddingTop: 3,
-				// With top and bottom both pinned, content-box would add the padding
-				// on top of the resolved height and push the block past the wrapper
-				// it spans.
-				boxSizing: 'border-box',
-				fontSize: 9,
-				color: GUI_THEME.dim2,
-				whiteSpace: 'nowrap',
-				// The label is wider than its block at month and year scopes; only
-				// one segment is ever highlighted, so it has nothing to overlap.
-				overflow: 'visible',
-			}}
-		>
-			{segment.label}
-		</div>
-	);
-};
-
-// The stretch a range drag has covered so far. Spans both charts and the gap,
-// like the segment highlight, because the window it will zoom to is one window
-// over both series.
-export const RangeSelection = ({from, to}: {from: number; to: number}) => {
-	const left = Math.min(from, to);
-
-	return (
-		<div
-			data-testid="scrubber-range-selection"
-			style={{
-				position: 'absolute',
-				left: `${left * 100}%`,
-				width: `${Math.abs(to - from) * 100}%`,
-				top: 0,
-				bottom: 0,
-				background: RANGE_SELECTION_COLOR,
-				borderLeft: `1px solid ${RANGE_SELECTION_EDGE}`,
-				borderRight: `1px solid ${RANGE_SELECTION_EDGE}`,
-				boxSizing: 'border-box',
-				pointerEvents: 'none',
-				// Over the needle: the needle marks where the board is, the selection
-				// what is about to replace the whole window.
-				zIndex: 4,
-			}}
-		/>
-	);
-};
-
-// The baseline every series is drawn against.
-export const TrackBaseline = ({
-	color,
-	anchor,
-}: {
-	color: string;
-	anchor: 'centre' | 'bottom' | 'top';
-}) => (
-	<div
-		style={{
-			position: 'absolute',
-			left: 0,
-			right: 0,
-			...(anchor === 'bottom' ? {bottom: 0} : {}),
-			...(anchor === 'top' ? {top: 0} : {}),
-			height: anchor === 'top' ? 1 : 2,
-			borderRadius: 999,
-			background: color,
-			// The centre line runs through the scatter's dots, so it sits a shade
-			// fainter than the edge-anchored baselines.
-			opacity: anchor === 'centre' ? 0.14 : 0.2,
-		}}
-	/>
-);
-
-// Absolute positioning ignores the track's padding, so top and bottom add it
-// back to line up with the points.
-export const HourAxisLabels = () => (
-	<>
-		{(
-			[
-				['00:00', {top: EVENTS_MODE_VERTICAL_PADDING}],
-				['12:00', {top: '50%', transform: 'translateY(-50%)'}],
-				['24:00', {bottom: EVENTS_MODE_VERTICAL_PADDING}],
-			] as const
-		).map(([label, position]) => (
-			<span
-				key={label}
-				style={{
-					position: 'absolute',
-					left: 2,
-					fontSize: 9,
-					color: GUI_THEME.dim,
-					pointerEvents: 'none',
-					...position,
-				}}
-			>
-				{label}
-			</span>
-		))}
-	</>
-);
-
-// The wrapper the entrance fade belongs on, never the individual bars or dots:
-// those are keyed by bucket time, so a scope change remounts each one and the
-// fade restarts per element as a full-chart flash. Callers key it per series —
-// the issue and commit layers are siblings, so a shared key collides and leaves
-// stale marks on screen.
-export const SeriesLayer = ({
-	animate,
-	children,
-}: {
-	animate: boolean;
-	children: React.ReactNode;
-}) => (
-	<div
-		style={{
-			position: 'absolute',
-			inset: 0,
-			pointerEvents: 'none',
-			animation: animate ? FADE_IN_ANIMATION : undefined,
-			// Walls this subtree off from the needle and highlight moving above
-			// it: without containment every frame of a drag re-lays-out the
-			// thousands of dots inside, and the drag stutters.
-			contain: 'layout paint',
-		}}
-	>
-		{children}
-	</div>
-);
-
-// Shared by the issue histogram and the commit one mirrored below it, so the
-// two halves cannot drift apart. Empty buckets draw nothing and stay hoverable
-// anyway, since hover is resolved arithmetically rather than by hit target.
-// Spans the full track height rather than the bar's, so empty buckets highlight
-// too. A sibling of the bars rather than one of them: it moves on every mouse
-// move, and the bars — up to MAX_TIME_BUCKETS nodes — must not re-render with
-// it.
-export const BucketHighlight = ({
-	index,
-	bucketCount,
-}: {
-	index: number;
-	bucketCount: number;
-}) => (
-	<div
-		style={{
-			position: 'absolute',
-			left: `${index * (100 / bucketCount)}%`,
-			top: 0,
-			bottom: 0,
-			width: `${100 / bucketCount}%`,
-			// A bucket can be thinner than a pixel at wide spans.
-			minWidth: 2,
-			background: BUCKET_HIGHLIGHT_COLOR,
-			pointerEvents: 'none',
-		}}
-	/>
-);
-
-const VolumeBarsImpl = ({
-	bars,
-	bucketCount,
-	firstBar,
-	lastBar,
-	color,
-	direction,
-	animate,
-}: {
-	bars: VolumeBar[];
-	bucketCount: number;
-	// The populated span, so the growth sweep runs across drawn bars only.
-	firstBar: number;
-	lastBar: number;
-	color: string;
-	// "up" is the issue histogram, "down" the mirrored commit one.
-	direction: 'up' | 'down';
-	animate: boolean;
-}) => {
-	const widthPercent = 100 / bucketCount;
-	const barWidth = barWidthCss(bucketCount);
-	const growsUp = direction === 'up';
-
-	// Only the bars present when this layer mounted are part of the entrance.
-	// A bar that appears later belongs to data arriving on its own, and must
-	// take its place without announcing itself.
-	const [fresh, setFresh] = useState(true);
-
-	useEffect(() => {
-		const timeout = setTimeout(() => setFresh(false), BAR_ENTRANCE_TOTAL_MS);
-		return () => clearTimeout(timeout);
-	}, []);
-
-	return (
-		<>
-			{bars.map(({index, intensity}) => (
-				<div
-					key={index}
-					style={{
-						position: 'absolute',
-						left: `${index * widthPercent}%`,
-						...(growsUp ? {bottom: 0} : {top: 0}),
-						width: barWidth,
-						height: 3 + intensity * (TRACK_HEIGHT - 3),
-						borderRadius: growsUp ? '1px 1px 0 0' : '0 0 1px 1px',
-						background: color,
-						opacity: 0.35 + intensity * 0.65,
-						transformOrigin: growsUp ? 'bottom' : 'top',
-						animation:
-							animate && fresh
-								? barGrowAnimation(index, firstBar, lastBar)
-								: undefined,
-						pointerEvents: 'none',
-					}}
-				/>
-			))}
-		</>
-	);
-};
-
-// Memoized because the track re-renders on every mouse move while the bars
-// themselves change only when the window or the filter does.
-export const VolumeBars = memo(VolumeBarsImpl);
-
-// Shared by both scatter series for the same reason. The vertical padding has
-// to be added back by hand: absolute positioning ignores the track's own.
-export const ScatterDot = ({
-	fraction,
-	hourFraction,
-	size,
-	color,
-	opacity,
-	zIndex,
-	title,
-	animation,
-	interactive = true,
-	onMouseEnter,
-	onMouseLeave,
-	onClick,
-}: {
-	// x is elapsed time across the window, y is time of day.
-	fraction: number;
-	hourFraction: number;
-	size: number;
-	color: string;
-	// Capped below 1 by callers so overlapping points blend rather than one
-	// hiding another.
-	opacity: number;
-	zIndex: number;
-	title: string;
-	animation: string | undefined;
-	// False while the series retracts, so a dot on its way out cannot be
-	// hovered for a hint or clicked into a diff.
-	interactive?: boolean;
-	onMouseEnter: () => void;
-	onMouseLeave: () => void;
-	onClick?: () => void;
-}) => (
-	<div
-		title={title}
-		// Without stopping it here, clicking a dot to inspect its diff would also
-		// start a scrub-drag on the track underneath.
-		onPointerDown={onClick && (event => event.stopPropagation())}
-		onClick={
-			onClick &&
-			(event => {
-				event.stopPropagation();
-				onClick();
-			})
-		}
-		onMouseEnter={onMouseEnter}
-		onMouseLeave={onMouseLeave}
-		style={{
-			position: 'absolute',
-			left: `${fraction * 100}%`,
-			top: EVENTS_MODE_VERTICAL_PADDING + hourFraction * EVENTS_SCATTER_HEIGHT,
-			width: size,
-			height: size,
-			borderRadius: '50%',
-			background: color,
-			opacity,
-			zIndex,
-			transform: `translate(${-size / 2}px, -50%)`,
-			animation,
-			pointerEvents: interactive ? 'auto' : 'none',
-			cursor: onClick ? 'pointer' : undefined,
-		}}
-	/>
-);
-
-// Lives on the chart wrapper rather than inside either chart, so it runs
-// unbroken through both and the gap between them.
-export const ScrubberNeedle = ({
-	fraction,
-	onGrab,
-}: {
-	fraction: number;
-	// A press here is a scrub, not the start of a range drag. The event still
-	// bubbles to the track, which owns the pointer capture and the moving.
-	onGrab: () => void;
-}) => {
-	const [hovered, setHovered] = useState(false);
-
-	const grip = {
-		onMouseEnter: () => setHovered(true),
-		onMouseLeave: () => setHovered(false),
-		onPointerDown: onGrab,
-	};
-
-	return (
-		<>
-			{/* A hairline is a 1px drag target. This is the same line's worth of
-			    grabbable width, invisible, centred on it. */}
-			<div
-				{...grip}
-				data-testid="scrubber-needle-grip"
-				style={{
-					position: 'absolute',
-					left: `${fraction * 100}%`,
-					top: 0,
-					bottom: 0,
-					width: NEEDLE_GRIP_WIDTH,
-					transform: `translateX(${-NEEDLE_GRIP_WIDTH / 2}px)`,
-					zIndex: 3,
-					pointerEvents: 'auto',
-					cursor: 'ew-resize',
-				}}
-			/>
-			<div
-				style={{
-					position: 'absolute',
-					left: `${fraction * 100}%`,
-					top: 0,
-					bottom: 0,
-					// A hairline, no glow: the needle marks an exact instant, and a
-					// soft edge blooms over bars that can be ~2px wide.
-					width: 1,
-					background: NEEDLE_COLOR,
-					zIndex: 3,
-					transform: 'translateX(-0.5px)',
-					pointerEvents: 'none',
-				}}
-			/>
-			<div
-				{...grip}
-				style={{
-					position: 'absolute',
-					left: `${fraction * 100}%`,
-					top: hovered ? -9 : -7,
-					width: 0,
-					height: 0,
-					borderLeft: `${hovered ? 6 : 5}px solid transparent`,
-					borderRight: `${hovered ? 6 : 5}px solid transparent`,
-					borderTop: `${hovered ? 8 : 7}px solid ${NEEDLE_COLOR}`,
-					zIndex: 3,
-					transform: `translateX(${hovered ? -6 : -5}px)`,
-					pointerEvents: 'auto',
-					cursor: 'ew-resize',
-				}}
-			/>
-		</>
-	);
-};
-
-// Hints hang below the charts; floating above covers the scope/mode controls,
-// which is the context being read while pointing at something.
-export const ScrubberHoverHint = ({
-	label,
-	rows,
-	segmentLabel,
-	stripeColor,
-	fraction,
-	trackWidthPx,
-	empty = false,
-}: {
-	label: string;
-	rows: string[];
-	// Drawn above the hint's own label, so the hovered moment names its period.
-	segmentLabel: string | undefined;
-	stripeColor: string;
-	// Fraction of the track the hint points at, and the track's measured width.
-	fraction: number;
-	trackWidthPx: number;
-	// An interval containing nothing. Still shown, but a shade down.
-	empty?: boolean;
-}) => (
-	<div
-		style={{
-			position: 'absolute',
-			top: '100%',
-			marginTop: 6,
-			left: clamp(
-				fraction * trackWidthPx - HOVER_HINT_WIDTH / 2,
-				0,
-				Math.max(0, trackWidthPx - HOVER_HINT_WIDTH),
-			),
-			width: HOVER_HINT_WIDTH,
-			// Border-box so `width` matches what the clamp math assumes; otherwise
-			// border and padding push the box past the track's edge.
-			boxSizing: 'border-box',
-			display: 'flex',
-			flexDirection: 'column',
-			gap: 2,
-			textAlign: 'left',
-			background: GUI_THEME.panel,
-			border: `1px solid ${GUI_THEME.line}`,
-			borderLeft: `3px solid ${empty ? GUI_THEME.dim : stripeColor}`,
-			// Smaller than the 6px used elsewhere: a rounder corner clips the 3px
-			// stripe into a visible wedge.
-			borderRadius: 3,
-			padding: '6px 10px',
-			pointerEvents: 'none',
-			// Above the board content this overhangs.
-			zIndex: 5,
-		}}
-	>
-		{segmentLabel && (
-			<div style={{fontSize: 10, color: GUI_THEME.dim}}>{segmentLabel}</div>
-		)}
-		<div
-			style={{
-				fontSize: 11,
-				fontWeight: 600,
-				color: empty ? GUI_THEME.secondary : GUI_THEME.primary,
-				whiteSpace: 'normal',
-				wordBreak: 'break-word',
-			}}
-		>
-			{label}
-		</div>
-		{rows.map((row, index) => (
-			<div
-				key={index}
-				style={{
-					fontSize: 11,
-					color: empty ? GUI_THEME.dim : GUI_THEME.secondary,
-					whiteSpace: 'normal',
-					wordBreak: 'break-word',
-				}}
-			>
-				{row}
-			</div>
-		))}
-	</div>
-);
-
-export type ScatterPoint = {
-	key: string;
-	// Set only on a per-event dot, which is what a log row can be matched to.
-	id: string | null;
-	// The moment the point stands for, for the hint's own labelling.
-	t: number;
-	// x across the window, y as time of day — the punchcard's two axes.
-	fraction: number;
-	hourFraction: number;
-	radius: number;
-	color: string;
-	opacity: number;
-	title: string;
-	commitSha: string | null;
-};
-
-// One series. Each animates on its own, so unticking "Code" retracts the
-// commits while the board events stay put.
-export type ScatterLayer = {
-	id: string;
-	points: ScatterPoint[];
-	// Changing this replays the entrance for this layer.
-	generation: string;
-	// On its way out: draw the retraction, then stop drawing it at all.
-	leaving: boolean;
-};
-
-type Phase = {mode: 'in' | 'out'; startedAt: number};
-
-// How near the pointer has to be, in px, to count as over a dot. Larger than
-// the dots themselves: they are 2px radius and would be almost unhittable.
-const HIT_RADIUS = 7;
-
-const pointAt = (
-	layers: readonly ScatterLayer[],
-	x: number,
-	y: number,
-	width: number,
-): ScatterPoint | null => {
-	let best: ScatterPoint | null = null;
-	let bestDistance = HIT_RADIUS * HIT_RADIUS;
-
-	for (const layer of layers) {
-		if (layer.leaving) continue;
-
-		for (const point of layer.points) {
-			const dx = point.fraction * width - x;
-			const dy =
-				EVENTS_MODE_VERTICAL_PADDING +
-				point.hourFraction * EVENTS_SCATTER_HEIGHT -
-				y;
-			const distance = dx * dx + dy * dy;
-
-			if (distance <= bestDistance) {
-				bestDistance = distance;
-				best = point;
-			}
-		}
-	}
-
-	return best;
-};
-
-// One node instead of one per event. At 2.2k dots the DOM version was 93% of
-// the whole document, and every frame that touched an ancestor paid for it.
-// Drawing them costs the same whether there are ten or ten thousand.
-// Everything else fades to this while one event is singled out, so the dot that
-// stays at full strength reads as the answer.
-const DIMMED_ALPHA = 0.08;
-const HIGHLIGHT_RADIUS_SCALE = 2.2;
-
-export const ScatterCanvas = ({
-	layers,
-	animate,
-	highlightId,
-	onPointEnter,
-	onPointLeave,
-	onPressCommit,
-}: {
-	layers: readonly ScatterLayer[];
-	animate: boolean;
-	// The one event to single out, or null for the plain chart.
-	highlightId: string | null;
-	onPointEnter: (point: ScatterPoint) => void;
-	onPointLeave: () => void;
-	// The commit a press landed on, or null for anywhere else. Reported rather
-	// than acted on: the track holds the pointer capture, so it is the one that
-	// can tell this press apart from the start of a range drag.
-	onPressCommit: (sha: string | null) => void;
-}) => {
-	const canvasRef = useRef<HTMLCanvasElement | null>(null);
-	const sizeRef = useRef({width: 0, height: 0});
-	// Read through refs by the draw loop, so a data change never restarts it.
-	const layersRef = useRef(layers);
-	layersRef.current = layers;
-	// Only dims when the highlighted event is actually on the chart. It may not
-	// be: an event outside the fetched window has no dot, and the bucketed
-	// fallback dots carry no id at all.
-	// Memoized: this walks every plotted point, and the track re-renders on
-	// every mouse move.
-	const activeHighlight = useMemo(
-		() =>
-			highlightId !== null &&
-			layers.some(layer => layer.points.some(point => point.id === highlightId))
-				? highlightId
-				: null,
-		[layers, highlightId],
-	);
-	const highlightRef = useRef(activeHighlight);
-	highlightRef.current = activeHighlight;
-	const phasesRef = useRef(new Map<string, Phase>());
-	const frameRef = useRef<number | null>(null);
-	const hoveredRef = useRef<string | null>(null);
-	// The scatter's entrance is drawn, not animated by CSS, so it emits no
-	// animationstart for a test to observe. This says the same thing.
-	const [entrancePlaying, setEntrancePlaying] = useState(false);
-
-	const paint = useCallback((now: number) => {
-		const canvas = canvasRef.current;
-		const context = canvas?.getContext('2d');
-		if (!canvas || !context) return false;
-
-		const {width, height} = sizeRef.current;
-		context.clearRect(0, 0, width, height);
-
-		let running = false;
-
-		for (const layer of layersRef.current) {
-			const phase = phasesRef.current.get(layer.id);
-			const elapsed = phase ? now - phase.startedAt : null;
-			const done = elapsed === null || elapsed >= DOT_EXIT_TOTAL_MS;
-
-			if (phase && !done) running = true;
-			// A finished exit leaves nothing behind.
-			if (phase?.mode === 'out' && done) continue;
-
-			for (const point of layer.points) {
-				const scale =
-					phase === null || phase === undefined || done
-						? 1
-						: phase.mode === 'in'
-						? dotEntranceScale(point.key, elapsed!)
-						: dotExitScale(point.key, elapsed!);
-
-				if (scale <= 0) continue;
-
-				const highlight = highlightRef.current;
-				const isHighlighted = highlight !== null && point.id === highlight;
-				const radiusScale =
-					scale * (isHighlighted ? HIGHLIGHT_RADIUS_SCALE : 1);
-				const cx = point.fraction * width;
-				const cy =
-					EVENTS_MODE_VERTICAL_PADDING +
-					point.hourFraction * EVENTS_SCATTER_HEIGHT;
-
-				context.globalAlpha =
-					highlight === null ? point.opacity : isHighlighted ? 1 : DIMMED_ALPHA;
-				context.fillStyle = point.color;
-				context.beginPath();
-				context.arc(cx, cy, point.radius * radiusScale, 0, Math.PI * 2);
-				context.fill();
-
-				// A halo as well as a bigger dot: at 2-3px a size change alone is easy
-				// to miss on a track holding hundreds of them.
-				if (isHighlighted) {
-					context.globalAlpha = 0.25;
-					context.beginPath();
-					context.arc(cx, cy, point.radius * radiusScale * 2.4, 0, Math.PI * 2);
-					context.fill();
-				}
-			}
-		}
-
-		context.globalAlpha = 1;
-
-		return running;
-	}, []);
-
-	const run = useCallback(() => {
-		if (frameRef.current !== null) return;
-
-		const step = () => {
-			frameRef.current = null;
-			const running = paint(performance.now());
-
-			if (running) frameRef.current = requestAnimationFrame(step);
-			else setEntrancePlaying(false);
-		};
-
-		frameRef.current = requestAnimationFrame(step);
-	}, [paint]);
-
-	// The backing store is sized in device pixels and the context scaled to
-	// match, or the dots are blurry on a retina display.
-	useEffect(() => {
-		const canvas = canvasRef.current;
-		const parent = canvas?.parentElement;
-		if (!canvas || !parent) return;
-
-		const resize = () => {
-			const ratio = window.devicePixelRatio || 1;
-			const {width, height} = parent.getBoundingClientRect();
-
-			sizeRef.current = {width, height};
-			canvas.width = Math.round(width * ratio);
-			canvas.height = Math.round(height * ratio);
-			canvas.style.width = `${width}px`;
-			canvas.style.height = `${height}px`;
-			canvas.getContext('2d')?.setTransform(ratio, 0, 0, ratio, 0, 0);
-			paint(performance.now());
-		};
-
-		resize();
-
-		const observer = new ResizeObserver(resize);
-		observer.observe(parent);
-
-		return () => observer.disconnect();
-	}, [paint]);
-
-	// Each layer starts its own entrance or exit. Scrubbing changes neither, so
-	// it never animates.
-	const signature = layers
-		.map(layer => `${layer.id}:${layer.generation}:${layer.leaving}`)
-		.join('|');
-
-	useEffect(() => {
-		const seen = new Set<string>();
-
-		for (const layer of layers) {
-			seen.add(layer.id);
-			const key = `${layer.generation}:${layer.leaving}`;
-			const marker = `${layer.id}@${key}`;
-			const current = phasesRef.current.get(layer.id);
-
-			if (
-				(current as (Phase & {marker?: string}) | undefined)?.marker === marker
-			)
-				continue;
-
-			if (!animate) {
-				phasesRef.current.delete(layer.id);
-				continue;
-			}
-
-			phasesRef.current.set(layer.id, {
-				mode: layer.leaving ? 'out' : 'in',
-				startedAt: performance.now(),
-				marker,
-			} as Phase);
-
-			if (!layer.leaving) setEntrancePlaying(true);
-		}
-
-		for (const id of [...phasesRef.current.keys()])
-			if (!seen.has(id)) phasesRef.current.delete(id);
-
-		if (animate) run();
-		else paint(performance.now());
-	}, [signature, animate, run, paint]);
-
-	// Repaint when the data or the highlight changes without a new entrance.
-	useEffect(() => {
-		if (frameRef.current === null) paint(performance.now());
-	}, [layers, activeHighlight, paint]);
-
-	useEffect(
-		() => () => {
-			if (frameRef.current !== null) cancelAnimationFrame(frameRef.current);
-		},
-		[],
-	);
-
-	const hitTest = (event: React.MouseEvent<HTMLCanvasElement>) => {
-		const rect = event.currentTarget.getBoundingClientRect();
-
-		return pointAt(
-			layersRef.current,
-			event.clientX - rect.left,
-			event.clientY - rect.top,
-			rect.width,
-		);
-	};
-
-	return (
-		<canvas
-			ref={canvasRef}
-			data-testid="scatter-canvas"
-			data-entrance={entrancePlaying ? 'playing' : 'done'}
-			data-highlight={activeHighlight ?? ''}
-			style={{position: 'absolute', inset: 0}}
-			onMouseMove={event => {
-				const point = hitTest(event);
-				if (point?.key === hoveredRef.current) return;
-
-				hoveredRef.current = point?.key ?? null;
-				if (point) onPointEnter(point);
-				else onPointLeave();
-			}}
-			onMouseLeave={() => {
-				hoveredRef.current = null;
-				onPointLeave();
-			}}
-			// Never stopped, not even over a commit: every press has to reach the
-			// track, or a range drag that begins on a dot never begins at all — and
-			// on a busy scatter most of them do. What the press turns out to mean is
-			// settled on release, by how far it travelled.
-			//
-			// There is deliberately no onClick here. The track takes pointer
-			// capture, which retargets the compatibility mouse events with it, so a
-			// click on the canvas is not the canvas's to hear.
-			onPointerDown={event => onPressCommit(hitTest(event)?.commitSha ?? null)}
-		/>
-	);
-};
